feat: Add GraphQL fragments
