+++ pages/index.js
@@ -3,10 +3,7 @@ import useSWR from 'swr'
 import cx from 'classnames'
 import format from 'date-fns/format'
 
-import {
-  AvailableProgramsQuery,
-  AvailableSampleProgramsQuery
-} from '@/queries/programs'
+import { ProgramsListQuery, SampleProgramsListQuery } from '@/queries/program'
 import Badge from '@/components/badge'
 import graphCmsClient from '@/lib/graphcms'
 import { getProduct } from '@/lib/db-admin'
@@ -31,14 +28,14 @@ function Index({ product }) {
     user
       ? hasSubscription
         ? [
-            AvailableProgramsQuery,
+            ProgramsListQuery,
             activeCategory,
             pagination.limit,
             pagination.offset,
             user.accessDate
           ]
         : [
-            AvailableSampleProgramsQuery,
+            SampleProgramsListQuery,
             activeCategory,
             pagination.limit,
             pagination.offset
